reorganize rules
